Base classes for distributions, with two example distributions.
Defined a couple of classes that basically serve as interfaces that can represent any kind of distribution. The idea is that we can have two kinds of distributions:

  1. Arbitrary distributions where all we can do is sample.
  2. Finite distributions that we can render as a PDF or CDF table.

This is captured in two abstract base classes: `Distribution` for 1 and `FiniteDistribution` for 2. Continuous distributions—which I have not implemented yet—would be subclasses of `Distribution` but *not* `FiniteDistribution`.

Having classes like this gives us a lot of flexibility: we have an explicit `Distribution` type to use in our process classes (MP, MRP, MDP... etc) and we can implement a lot of *different* distributions using different technologies under the hood (ie our own logic vs calling numpy vs whatever else) in a way that hides those details and lets us use different implementations totally interchangeably in the rest of our code.

This is a key software design decision that we can talk about in the book: by using interfaces, we ensure that our process code and algorithms only need to specify what functionality they *need* from a distribution (ie tabular vs sampling), without needing to know exactly how the distribution is computed. This cleanly separates the concerns between different parts of the code and makes it easy to add both new sampling algorithms *and* new optimization algorithms without either needing to know about the other.

To illustrate how this works, I added two simple example distributions:

  1. A `Bernoulli` distribution that always returns booleans.
  2. A `Choose` distribution that choose an outcome from a list uniformly at random.

As we go along, we can define more distributions in this style.