Refactors DP code to return iterators
The idea is that the caller of each algorithm can decide when and how the algorithm should converge. This means we don't have to hard-code a convergence criterion and callers of the function can access intermediate steps in the algorithm (great for debugging, visualization... etc).

I also fixed up most of the mypy errors I ran into, which resulted in a few extra if ... is None style checks to account for terminal states.