Refactored FiniteDistribution to use dicts instead of lists
As I was playing around with some new finite horizon code, I found that the existing `FiniteDistribution` design was a bit awkward to work with. `FiniteDistribution` used `Sequence[Tuple[A, float]]` to represent distributions in a tabular format, but it's substantially more convenient to use `Mapping[A, float]` for the same task instead.

I believe that I originally went with the `Sequence` design so that we could build distributions of objects that don't work as dictionary keys (like Python functions). However, the code we've written so far doesn't use that and using dicts for distributions ends up a lot neater than using lists of pairs, so I figure we should just use dicts and then deal with non-hashable values if we ever get to needing them.

This refactoring does a few other things to improve the distribution code:

  1. Make `FiniteDistribution` iterable. Now you can write `for x, p in dist:` instead of needing `dist.table()`.
  2. Implement reasonable `__eq__` and `__repr__` methods for `FiniteDistribution`.
  3. Implement a `map` method for `FiniteDistribution` so that we can apply a function to each outcome.
  4. Add a simple test suite for the distribution code to help catch bugs in the future. This will make refactoring this code again in the future easier.

These changes probably require changing existing code that uses distributions. I changed everything in the `rl` directory—or at least everything mypy pointed out—but haven't looked anywhere else. In practice, I expect most places you called `.table()` need to be changed, probably by just removing the call.