Added .pylintrc and .pycheckers.
`.pylintrc` is a config file for `pylint`, a tool I've been using to spot and fix little Python issues. I recommend giving it a try.

`.pycheckers` is a config file for a tool I'm using in Emacs to run multiple checkers (ie pylint + mypy). It shouldn't have any effect for other editor setups.