Organized code into Python package with a setup.cfg file.
This organization means that we can install our framework code as a package with pip. This isn't super important right now, but it helps with my local development tools.

For local development, we can install the package in "editable" mode from the top-level directory of the repo:

```
pip install --editable .
```