Added an is_terminal method to MarkovDecisionProcess.
Since `MarkovProcess` has an `is_terminal` method, having an analogous method for `MarkovDecisionsProcess` makes sense—good to keep our APIs consistent in design.

The default implementation of `is_terminal(s)` checks if `actions(s)` returns an emtpy result. This can be overridden by concrete implementations of `MarkovDecisionProcess` if there are more direct and efficient ways to check whether a state is terminal.

The upside to this design is that `is_terminal` makes checking for terminal states cleaner and more explicit. The downside is that an overridden `is_terminal` could be inconsistent with `actions`, but that is an easy property to test for.