Added support for grouping chapters into sections.
With this change, the books is now divided first into sections (I, II, III... etc) and then into chapters. To support this, I had to increase the level of each header by one. Before this change, a chapter looked like this:

```

... etc ...
```

Now it looks like this:

```

... etc ...
```

The first chapter of each section (like Chapter 0 and Chapter 6) have a top-level header for the section:

```

... etc ...
```

This approach of having extra levels of headers feels like a bit of  a hack, but it seems to work well enough. We can figure out something a bit better if we run into problems in the future.