Fix seed to make tests deterministic.
Our TD and MC tests are randomly failing because the bounds are a bit too tight. Instead of constantly fiddling with the settings and bounds for those tests, I fixed the random seed in the two modules so that they always run the same way.

This will keep our PRs from randomly failing CI.