Fixed tests for Python 3.7
We've been developing with Python 3.8 rather than 3.7, and under 3.7 our tests failed with the following error in a few places:

```
>       return itertools.accumulate(transitions, step, initial=approx_0)
E       TypeError: accumulate() takes at most 2 arguments (3 given)
```

Turns out the problem is that the `initial` argument to `itertools.accumulate` was only added in 3.8. Since we use this argument extensively in both the library code and book examples, I solved this problem by defining a version of `accumulate` in `rl/iterate.py` that has an `initial` argument:

```python
def accumulate(
        iterable: Iterable[X],
        func: Callable[[Y, X], Y],
        *,
        initial: Optional[Y]
) -> Iterator[Y]:
    # ...
```

I then replaced uses of `itertools.accumulate` in both the library code and book examples with `iterate.accumulate`:

```python
import rl.iterate as iterate
...
def evaluate_mrp(...):
    ...
    return iterate.accumulate(transitions, step, initial=approx_0)
```

I updated the text in a couple of places to add a note about this, but not sure that's the best approach. For example,

> Note the use of the standard Python function [`itertools.accumulate`](https://docs.python.org/3/library/itertools.html#itertools.accumulate)...

became:

> Note the use of the `rl.iterate.accumulate` function (a wrapped version of [`itertools.accumulate`](https://docs.python.org/3/library/itertools.html#itertools.accumulate))...

After this change, our test suite passed with Python 3.7. To catch these issues in the future, I added some code to `default.nix` that lets us open shells with Python 3.6, 3.7, 3.8 or 3.9:

```sh
nix-shell --argstr python-version '3.7'
```

I also made `3.7` the default version of Python that runs tests on every PR in this repo.