Reorganized book files + updated to-pdf script
To make the repo a bit easier to navigate, I moved all the directories with book content (chapters, appendices... etc) into a single `book` directory and updated `bin/to-pdf` to handle that correctly.

I also created a new file called `structure` that lists the chapters in order. This seems a bit more readable than having the chapters hardcoded directly in the `to-pdf` script.