Fix missing Deja Vu font issue on macOS
Turns out that we can't reference custom fonts by name on macOS. I switched the monospace font for the book to use Deja Vu Sans Mono which is automatically installed in the Nix shell and it worked on Linux, but ran into this problem on macOS.

To fix the problem, we need to refer to the font by its filepath, not its name. I did this having the Nix shell set `DEJA_VU_DIRECTORY` to the path of the installed Deja Vu fonts, then did some TeX trickery to have the generated LaTeX file read this variable and fetch the font correctly.

@coverdrive, could you give this a try on this branch? The new setup worked on my Linux machine, but I haven't tested whether this actually solves the problem on macOS. Ideally, the PDF should build exactly the same as before (running `bin/to-pdf` from a `nix-shell`) with no changes needed on your end.