Change the page dimensions and text size back for the non-TF format
When I added support for the custom TF CLS file, I had to disable the `geometry` package for it to work. However, this change made the default rendering (with the `scrbook` class) revert back to its normal margin sizes, which don't look great on PDF. This PR adds the `geometry` package settings back to the LaTeX template, but only if the `tf-format` variable isn't set.

I had also changed the default text size to 10pt, but since there's no point in having smaller text in the free PDF version of the book, I set this back to 12pt.