Remove Travis CI
We've run out of credits on Travis CI. Since we're not making too many changes to the library code right now, I'm just disabling the Travis integration altogether.

I'll move the Python tests over to GitHub Actions at some point, but for now we can just run them manually if we need to.