Only build PDFs on pushes to master
Removing the Travis config in #237, I noticed that the GitHub Actions to build the PDF were being run twice (push + pull request) and took a while.

We don't really need to build the book PDF that often, so I'm reconfiguring it to only build on pushes to `master` (ie when a PR gets merged).