Reduced the size of most of the figures
Previously, images would automatically be scaled up to fit the full width of the page, which was too large for most of them. I went through and resized the images to fit better with the size of the text and to fix layout issues (figures being floated into weird places). As a bonus, this also cuts down the page count a little bit.