Fixed hyphenation issues from em-dashes
TeX won't hyphenate words that are connected with `---` and no spaces (`like--this`), which was causing some lines in the generated PDF to go into the margin. To fix this, we can define and use `\emdash` instead of `---`:

``` latex
\newcommand{\emdash}{\nobreak---\nobreak\hskip0pt}
```

Instead of doing the substitution manually, I added a Pandoc filter that automatically replaces `---` and `—` with `\emdash ` when generating TeX. On a cursory inspection, this solves (most?) of the em-dash hyphenation issues without causing additional problems.